# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Memory Cloud API
  version: 1.0.0
  description: |
    Persistent memory infrastructure for AI agent fleets.

    The Agent Memory Cloud (AMC) provides a centralized memory system that allows
    AI agents to store, retrieve, and search memories across sessions and projects.

    ## Key Features
    - **Persistent Memory**: Store memories that survive across sessions
    - **Semantic Search**: Find relevant memories using natural language queries
    - **Agent & Project Organization**: Organize memories by agent and project
    - **Memory Types**: Support for different memory types (outcome, learning, preference, decision)
    - **Tagging & Metadata**: Flexible categorization system

    ## Authentication
    All API endpoints require authentication using either:
    - **API Key**: `X-API-Key` header
    - **JWT Token**: `Authorization: Bearer <token>` header

    ## Rate Limits
    - Standard tier: 100 requests/minute
    - Pro tier: 1000 requests/minute

    ## Pagination
    List endpoints support cursor-based pagination:
    - `limit`: Number of items per page (default: 20, max: 100)
    - `offset`: Offset for pagination (default: 0)

  contact:
    name: Faintech Solutions
    email: support@faintech.dev
    url: https://faintech.dev

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.faintech.dev/v1
    description: Production server
  - url: http://localhost:8000/v1
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...
paths:
...
  /auth/register:
    post:
      summary: Register user
      description: Create a new user account
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8

      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
  /auth/login:
    post:
      summary: Login
      description: Authenticate and get JWT tokens
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string

      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                    example: "bearer"
  /auth/refresh:
    post:
      summary: Refresh token
      description: Refresh access token using refresh token
      operationId: refreshToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string

      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
  /auth/logout:
    post:
      summary: Logout
      description: Logout and invalidate tokens
      operationId: logout
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logout successful